refactor(ios): split BackupSettingsView into support banner and iCloud section
